+++ base/precompile.jl
@@ -447,22 +447,10 @@ precompile(Tuple{typeof(Base.copy), Base.GenericIOBuffer{Array{UInt8, 1}}})
 precompile(Tuple{typeof(Base.getindex), Base.Dict{Symbol, Any}, Symbol})
 precompile(Tuple{typeof(Core.Compiler.length), Tuple{Core.Compiler.Const, Core.Compiler.Const, DataType}})
 precompile(Tuple{typeof(Core.Compiler.getindex), Tuple{Core.Compiler.Const, Core.Compiler.Const, DataType}, Int64})
-precompile(Tuple{typeof(Base.Docs.helpmode), String})
 precompile(Tuple{typeof(Base.ht_keyindex), Base.Dict{Symbol, Base.Docs.DocStr}, Symbol})
-precompile(Tuple{typeof(Base.Docs.helpmode), Base.TTY, String})
-precompile(Tuple{getfield(Base.Docs, Symbol("#@repl")), Base.TTY, Symbol})
-precompile(Tuple{typeof(Base.Docs.repl), Base.TTY, Symbol})
-precompile(Tuple{typeof(Base.Docs._repl), Symbol})
 precompile(Tuple{getfield(Core, Symbol("#@doc")), Symbol})
 precompile(Tuple{typeof(Base.Docs.docm), LineNumberNode, Module, Symbol})
-precompile(Tuple{typeof(Base._setindex!), Base.Dict{Any, Any}, Base.Markdown.Config, Symbol, Int64})
-precompile(Tuple{Type{Base.Markdown.MD}})
-precompile(Tuple{typeof(Base.setindex!), Base.Dict{Any, Any}, Base.Markdown.Config, Symbol})
-precompile(Tuple{typeof(Base.Markdown.skipblank), Base.GenericIOBuffer{Array{UInt8, 1}}})
 precompile(Tuple{typeof(Base.vcat), Array{Function, 1}, Array{Function, 1}})
-precompile(Tuple{typeof(Base.Docs.formatdoc), Base.Docs.DocStr})
-precompile(Tuple{typeof(Base.Docs.parsedoc), Base.Docs.DocStr})
-precompile(Tuple{typeof(Base.Docs.repl_latex), Base.TTY, String})
 precompile(Tuple{typeof(Base.ht_keyindex), Base.Dict{String, String}, String})
 precompile(Tuple{typeof(Base.ht_keyindex), Base.Dict{Symbol, Tuple{String, String}}, Symbol})
 precompile(Tuple{typeof(Base.print), Base.TTY, Char, Char})
@@ -471,17 +459,13 @@ precompile(Tuple{typeof(Base.ht_keyindex2!), Base.Dict{String, String}, String})
 precompile(Tuple{typeof(Base._setindex!), Base.Dict{String, String}, String, String, Int64})
 precompile(Tuple{typeof(Core.Compiler.isbits), Base.Dict{String, String}})
 precompile(Tuple{typeof(Base.setindex!), Base.Dict{String, String}, String, String})
-precompile(Tuple{typeof(Base.Docs.symbol_latex), String})
-precompile(Tuple{typeof(Base.Docs.repl_search), Base.TTY, String})
 precompile(Tuple{typeof(Base.print), Base.TTY, String})
 precompile(Tuple{typeof(Base.print), Base.TTY, Char})
 precompile(Tuple{typeof(Base.print), Base.TTY, String, Char})
 precompile(Tuple{typeof(Base.mapreduce_impl), typeof(Base.identity), typeof(Base.:(+)), Array{Int64, 1}, Int64, Int64, Int64})
 precompile(Tuple{typeof(Base._mapreduce), typeof(Base.identity), typeof(Base.:(+)), Base.IndexLinear, Array{Int64, 1}})
 precompile(Tuple{typeof(Base.popfirst!), Array{Char, 1}})
-precompile(Tuple{typeof(Base.Docs.avgdistance), Array{Int64, 1}})
 precompile(Tuple{typeof(Core.Compiler.isbits), Tuple{Array{String, 1}}})
-precompile(Tuple{typeof(Base.Docs.accessible), Module})
 precompile(Tuple{Type{Base.Generator{I, F} where F where I}, typeof(Base.names), Array{Any, 1}})
 precompile(Tuple{typeof(Base._collect), Array{Any, 1}, Base.Generator{Array{Any, 1}, typeof(Base.names)}, Base.EltypeUnknown, Base.HasShape})
 precompile(Tuple{typeof(Base.collect_to!), Array{Array{Symbol, 1}, 1}, Base.Generator{Array{Any, 1}, typeof(Base.names)}, Int64, Int64})
@@ -519,8 +503,6 @@ precompile(Tuple{Type{Base.Generator{I, F} where F where I}, typeof(Base.string)
 precompile(Tuple{typeof(Base._collect), Array{Any, 1}, Base.Generator{Array{Any, 1}, typeof(Base.string)}, Base.EltypeUnknown, Base.HasShape})
 precompile(Tuple{typeof(Base.similar), Array{Any, 1}, Type{String}, Tuple{Base.OneTo{Int64}}})
 precompile(Tuple{typeof(Base.collect_to!), Array{String, 1}, Base.Generator{Array{Any, 1}, typeof(Base.string)}, Int64, Int64})
-precompile(Tuple{getfield(Base.Docs, Symbol("#kw##matchinds")), Array{Any, 1}, typeof(Base.Docs.matchinds), String, String})
-precompile(Tuple{typeof(Base.Docs.bestmatch), String, String})
 precompile(Tuple{typeof(Base.length), Tuple{DataType, DataType}})
 precompile(Tuple{Type{BoundsError}, Array{Int64, 2}, Tuple{Base.UnitRange{Int64}, Int64}})
 precompile(Tuple{typeof(Base.throw_boundserror), Array{Int64, 2}, Tuple{Base.UnitRange{Int64}, Int64}})
@@ -560,8 +542,6 @@ precompile(Tuple{typeof(Base._unsafe_setindex!), Base.IndexLinear, Array{Int64, 
 precompile(Tuple{Type{BoundsError}, Array{Int64, 2}, Tuple{Int64, Base.UnitRange{Int64}}})
 precompile(Tuple{typeof(Base.throw_boundserror), Array{Int64, 2}, Tuple{Int64, Base.UnitRange{Int64}}})
 precompile(Tuple{typeof(Base._unsafe_setindex!), Base.IndexLinear, Array{Int64, 2}, Base.UnitRange{Int64}, Int64, Base.UnitRange{Int64}})
-precompile(Tuple{typeof(Base.Docs.levenshtein), String, String})
-precompile(Tuple{typeof(Base.Docs.fuzzyscore), String, String})
 precompile(Tuple{typeof(Core.Compiler.copy_exprargs), Array{Any, 1}})
 precompile(Tuple{typeof(Core.Compiler.copy), Expr})
 precompile(Tuple{typeof(Core.Compiler.copyto!), Array{Any, 1}, Core.Compiler.Generator{Array{Any, 1}, typeof(Core.Compiler.copy_exprs)}})
@@ -609,9 +589,6 @@ precompile(Tuple{typeof(Base.throw_checksize_error), Array{String, 1}, Tuple{Bas
 precompile(Tuple{typeof(Base._unsafe_getindex), Base.IndexLinear, Array{String, 1}, Array{Int64, 1}})
 precompile(Tuple{typeof(Base.reverse), Array{String, 1}})
 precompile(Tuple{typeof(Base.reverse), Array{String, 1}, Int64, Int64})
-precompile(Tuple{getfield(Base.Docs, Symbol("#kw##printmatches")), Array{Any, 1}, typeof(Base.Docs.printmatches), Base.TTY, String, Array{String, 1}})
-precompile(Tuple{typeof(Base.Docs.printmatch), Base.TTY, String, String})
-precompile(Tuple{typeof(Base.Docs.doc), Base.Docs.Binding})
 precompile(Tuple{typeof(Base.print), Base.GenericIOBuffer{Array{UInt8, 1}}, String, Base.Docs.Binding, String, String, String, Char})
 precompile(Tuple{typeof(Base.show_method_table), Base.GenericIOBuffer{Array{UInt8, 1}}, Base.MethodList, Int64, Bool})
 precompile(Tuple{typeof(Base.print), Base.GenericIOBuffer{Array{UInt8, 1}}, Base.MethodList})
@@ -624,7 +601,6 @@ precompile(Tuple{typeof(Base.Pkg.dir), String, String})
 precompile(Tuple{typeof(Base._setindex!), Base.Dict{Any, Any}, Array{Base.Docs.DocStr, 1}, Symbol, Int64})
 precompile(Tuple{typeof(Base._setindex!), Base.Dict{Any, Any}, Base.Docs.Binding, Symbol, Int64})
 precompile(Tuple{typeof(Base._setindex!), Base.Dict{Any, Any}, Type{Union{}}, Symbol, Int64})
-precompile(Tuple{typeof(Base._collect), Array{Base.Docs.DocStr, 1}, Base.Generator{Array{Base.Docs.DocStr, 1}, typeof(Base.Docs.parsedoc)}, Base.EltypeUnknown, Base.HasShape})
 precompile(Tuple{typeof(Base.ht_keyindex2!), Base.Dict{Union{DataType, typeof(Type)}, Nothing}, typeof(Type)})
 precompile(Tuple{typeof(Base._setindex!), Base.Dict{Union{DataType, typeof(Type)}, Nothing}, Nothing, typeof(Type), Int64})
 precompile(Tuple{typeof(Base.ht_keyindex2!), Base.Dict{Union{DataType, typeof(Type)}, Nothing}, DataType})
@@ -649,8 +625,6 @@ precompile(Tuple{typeof(Base.setindex!), Base.Dict{Any, Any}, Array{Base.Docs.Do
 precompile(Tuple{typeof(Base.setindex!), Base.Dict{Any, Any}, Base.Docs.Binding, Symbol})
 precompile(Tuple{typeof(Base.setindex!), Base.Dict{Any, Any}, Type{Union{}}, Symbol})
 precompile(Tuple{typeof(Base.Docs.aliasof), Base.Docs.Binding})
-precompile(Tuple{typeof(Base.Docs.summarize), Base.Docs.Binding, Type{Union{}}})
-precompile(Tuple{typeof(Base.Docs.doc), Base.Docs.Binding, Type{Union{}}})
 precompile(Tuple{typeof(Base.haskey), Base.IdDict, Base.Docs.Binding})
 precompile(Tuple{typeof(Base.getindex), Base.IdDict, Base.Docs.Binding})
 precompile(Tuple{typeof(Base.push!), Array{Base.Docs.MultiDoc, 1}, Base.Docs.MultiDoc})
@@ -659,43 +633,8 @@ precompile(Tuple{typeof(Base.done), Array{Type{T} where T, 1}, Int64})
 precompile(Tuple{typeof(Base.next), Array{Type{T} where T, 1}, Int64})
 precompile(Tuple{typeof(Base.getindex), Base.IdDict, Type{Tuple{Any}}})
 precompile(Tuple{typeof(Base.push!), Array{Base.Docs.DocStr, 1}, Base.Docs.DocStr})
-precompile(Tuple{typeof(Base.Docs.formatdoc), Base.GenericIOBuffer{Array{UInt8, 1}}, Base.Docs.DocStr, String})
-precompile(Tuple{typeof(Base.Markdown.hashheader), Base.GenericIOBuffer{Array{UInt8, 1}}, Base.Markdown.MD})
-precompile(Tuple{typeof(Base.Markdown.parseinline), Base.GenericIOBuffer{Array{UInt8, 1}}, Base.Markdown.MD, Array{Function, 1}})
-precompile(Tuple{typeof(Base.Markdown.config), Base.Markdown.MD})
-precompile(Tuple{typeof(Base.Markdown.parseinline), Base.GenericIOBuffer{Array{UInt8, 1}}, Base.Markdown.MD, Base.Markdown.Config})
-precompile(Tuple{typeof(Base.Markdown.list), Base.GenericIOBuffer{Array{UInt8, 1}}, Base.Markdown.MD})
-precompile(Tuple{typeof(Base._searchindex), Base.SubString{String}, String, Int64})
-precompile(Tuple{getfield(Base.Markdown, Symbol("#kw##skipwhitespace")), Array{Any, 1}, typeof(Base.Markdown.skipwhitespace), Base.GenericIOBuffer{Array{UInt8, 1}}})
-precompile(Tuple{typeof(Base.ht_keyindex), Base.Dict{Symbol, Base.Markdown.Config}, Symbol})
-precompile(Tuple{getfield(Base.Markdown, Symbol("#kw##parse")), Array{Any, 1}, typeof(Base.Markdown.parse), Base.GenericIOBuffer{Array{UInt8, 1}}})
-precompile(Tuple{getfield(Base.Markdown, Symbol("#kw##startswith")), Array{Any, 1}, typeof(Base.Markdown.startswith), Base.GenericIOBuffer{Array{UInt8, 1}}, Base.Regex})
-precompile(Tuple{typeof(Core.Compiler.isbits), Tuple{Int64, Base.Regex}})
-precompile(Tuple{typeof(Base.contains), String, String})
-precompile(Tuple{typeof(Base.contains), Base.SubString{String}, String})
-precompile(Tuple{typeof(Base.Markdown.pushitem!), Base.Markdown.List, Base.GenericIOBuffer{Array{UInt8, 1}}})
-precompile(Tuple{typeof(Base.Markdown.fencedcode), Base.GenericIOBuffer{Array{UInt8, 1}}, Base.Markdown.MD})
-precompile(Tuple{getfield(Base.Markdown, Symbol("#kw##startswith")), Array{Any, 1}, typeof(Base.Markdown.startswith), Base.GenericIOBuffer{Array{UInt8, 1}}, String})
-precompile(Tuple{typeof(Base.skip), Base.GenericIOBuffer{Array{UInt8, 1}}, Int64})
-precompile(Tuple{typeof(Base.lstrip), String, Char})
-precompile(Tuple{typeof(Base.Markdown.blockquote), Base.GenericIOBuffer{Array{UInt8, 1}}, Base.Markdown.MD})
-precompile(Tuple{getfield(Base.Markdown, Symbol("#kw##parse")), Array{Any, 1}, typeof(Base.Markdown.parse), String})
-precompile(Tuple{typeof(Base.Markdown.admonition), Base.GenericIOBuffer{Array{UInt8, 1}}, Base.Markdown.MD})
-precompile(Tuple{getfield(Base.Markdown, Symbol("#kw##linecontains")), Array{Any, 1}, typeof(Base.Markdown.linecontains), Base.GenericIOBuffer{Array{UInt8, 1}}, String})
-precompile(Tuple{typeof(Base.ucfirst), Base.SubString{String}})
-precompile(Tuple{typeof(Base.Markdown.blocktex), Base.GenericIOBuffer{Array{UInt8, 1}}, Base.Markdown.MD})
 precompile(Tuple{typeof(Core.Compiler.length), Tuple{DataType, Core.Compiler.Const, DataType}})
 precompile(Tuple{typeof(Core.Compiler.getindex), Tuple{DataType, Core.Compiler.Const, DataType}, Int64})
-precompile(Tuple{getfield(Base.Markdown, Symbol("#kw##parse_inline_wrapper")), Array{Any, 1}, typeof(Base.Markdown.parse_inline_wrapper), Base.GenericIOBuffer{Array{UInt8, 1}}, String})
-precompile(Tuple{typeof(Base.Markdown.startswith), Base.GenericIOBuffer{Array{UInt8, 1}}, String})
-precompile(Tuple{typeof(Base.Markdown.blockinterp), Base.GenericIOBuffer{Array{UInt8, 1}}, Base.Markdown.MD})
-precompile(Tuple{getfield(Base, Symbol("#kw##parse")), Array{Any, 1}, typeof(Base.parse), Base.GenericIOBuffer{Array{UInt8, 1}}})
-precompile(Tuple{typeof(Base.Markdown.interpinner), Base.GenericIOBuffer{Array{UInt8, 1}}, Bool})
-precompile(Tuple{typeof(Base.Markdown.indentcode), Base.GenericIOBuffer{Array{UInt8, 1}}, Base.Markdown.MD})
-precompile(Tuple{typeof(Base.Markdown.footnote), Base.GenericIOBuffer{Array{UInt8, 1}}, Base.Markdown.MD})
-precompile(Tuple{Type{Base.Markdown.Footnote}, Nothing, Array{Any, 1}})
-precompile(Tuple{Type{Base.Markdown.Footnote}, Base.SubString{String}, Array{Any, 1}})
-precompile(Tuple{typeof(Base.Markdown.github_table), Base.GenericIOBuffer{Array{UInt8, 1}}, Base.Markdown.MD})
 precompile(Tuple{typeof(Base.replace), String, Pair{String,String}})
 precompile(Tuple{typeof(Base.pop!), Array{Base.SubString{String}, 1}})
 precompile(Tuple{typeof(Base.union!), Base.Set{Char}, String})
@@ -705,54 +644,10 @@ precompile(Tuple{typeof(Core.Compiler.length), Tuple{Core.Compiler.Const, Core.C
 precompile(Tuple{typeof(Core.Compiler.getindex), Tuple{Core.Compiler.Const, Core.Compiler.Const, Core.Compiler.Const, Core.Compiler.Const, Core.Compiler.Const}, Int64})
 precompile(Tuple{typeof(Base.throw_boundserror), Base.SubArray{UInt8, 1, Array{UInt8, 1}, Tuple{Base.UnitRange{Int64}}, true}, Tuple{Int64}})
 precompile(Tuple{typeof(Base.peek), Base.GenericIOBuffer{Base.SubArray{UInt8, 1, Array{UInt8, 1}, Tuple{Base.UnitRange{Int64}}, true}}})
-precompile(Tuple{typeof(Base.Markdown.parseinline), Base.GenericIOBuffer{Base.SubArray{UInt8, 1, Array{UInt8, 1}, Tuple{Base.UnitRange{Int64}}, true}}, Base.Markdown.MD, Array{Function, 1}})
-precompile(Tuple{typeof(Base.read), Base.GenericIOBuffer{Base.SubArray{UInt8, 1, Array{UInt8, 1}, Tuple{Base.UnitRange{Int64}}, true}}, Type{Char}})
-precompile(Tuple{typeof(Base.Markdown.parseinline), Base.GenericIOBuffer{Base.SubArray{UInt8, 1, Array{UInt8, 1}, Tuple{Base.UnitRange{Int64}}, true}}, Base.Markdown.MD, Base.Markdown.Config})
-precompile(Tuple{typeof(Base.copyto!), Base.IndexLinear, Array{Array{Any, 1}, 1}, Base.IndexLinear, Array{Any, 1}})
-precompile(Tuple{typeof(Base.Markdown.parsealign), Nothing})
-precompile(Tuple{typeof(Base.Markdown.parsealign), Array{Base.SubString{String}, 1}})
-precompile(Tuple{typeof(Base.Markdown.rowlength!), Nothing, Int64})
-precompile(Tuple{typeof(Base.Markdown.rowlength!), Array{Base.SubString{String}, 1}, Int64})
-precompile(Tuple{Type{Base.Markdown.Table}, Array{Any, 1}, Nothing})
-precompile(Tuple{Type{Base.Markdown.Table}, Array{Any, 1}, Array{Symbol, 1}})
-precompile(Tuple{typeof(Base.Markdown.horizontalrule), Base.GenericIOBuffer{Array{UInt8, 1}}, Base.Markdown.MD})
-precompile(Tuple{typeof(Core.Compiler.isbits), Base.Markdown.HorizontalRule})
-precompile(Tuple{typeof(Base.Markdown.setextheader), Base.GenericIOBuffer{Array{UInt8, 1}}, Base.Markdown.MD})
-precompile(Tuple{typeof(Base.Markdown.paragraph), Base.GenericIOBuffer{Array{UInt8, 1}}, Base.Markdown.MD})
-precompile(Tuple{getfield(Base.Markdown, Symbol("#kw##parse")), Array{Any, 1}, typeof(Base.Markdown.parse), Base.GenericIOBuffer{Array{UInt8, 1}}, Base.Markdown.MD, Base.Markdown.Config})
-precompile(Tuple{getfield(Base.Markdown, Symbol("#kw##parse")), Array{Any, 1}, typeof(Base.Markdown.parse), Base.GenericIOBuffer{Array{UInt8, 1}}, Base.Markdown.MD})
-precompile(Tuple{typeof(Base.Markdown.inline_code), Base.GenericIOBuffer{Array{UInt8, 1}}, Base.Markdown.MD})
-precompile(Tuple{typeof(Base.Markdown.readuntil), Base.GenericIOBuffer{Array{UInt8, 1}}, String})
-precompile(Tuple{typeof(Base.Markdown.readuntil), Base.GenericIOBuffer{Array{UInt8, 1}}, Base.SubString{String}})
 precompile(Tuple{typeof(Base.setindex!), Base.Dict{Any, Any}, Module, Symbol})
 precompile(Tuple{typeof(Base._setindex!), Base.Dict{Any, Any}, Module, Symbol, Int64})
 precompile(Tuple{typeof(Base.setindex!), Base.Dict{Any, Any}, String, Symbol})
 precompile(Tuple{typeof(Base._setindex!), Base.Dict{Any, Any}, String, Symbol, Int64})
-precompile(Tuple{typeof(Base.similar), Array{Base.Docs.DocStr, 1}, Type{Base.Markdown.MD}, Tuple{Base.OneTo{Int64}}})
-precompile(Tuple{typeof(Base.setindex!), Array{Base.Markdown.MD, 1}, Base.Markdown.MD, Int64})
-precompile(Tuple{typeof(Base.collect_to!), Array{Base.Markdown.MD, 1}, Base.Generator{Array{Base.Docs.DocStr, 1}, typeof(Base.Docs.parsedoc)}, Int64, Int64})
-precompile(Tuple{typeof(Base.unsafe_copyto!), Array{Base.Markdown.MD, 1}, Int64, Array{Base.Markdown.MD, 1}, Int64, Int64})
-precompile(Tuple{typeof(Base.Docs.catdoc), Base.Markdown.MD})
-precompile(Tuple{typeof(Base.copyto!), Base.IndexLinear, Array{Any, 1}, Base.IndexLinear, Array{Base.Markdown.MD, 1}})
-precompile(Tuple{Type{Base.Markdown.MD}, Array{Base.Markdown.MD, 1}})
-precompile(Tuple{Type{Base.Markdown.MD}, Base.Markdown.MD})
-precompile(Tuple{typeof(Base.vcat), Base.Markdown.MD})
-precompile(Tuple{typeof(Base.put!), Base.Channel{Any}, Tuple{Base.Markdown.MD, Nothing}})
-precompile(Tuple{typeof(Base.put_buffered), Base.Channel{Any}, Tuple{Base.Markdown.MD, Nothing}})
-precompile(Tuple{typeof(Base.put_unbuffered), Base.Channel{Any}, Tuple{Base.Markdown.MD, Nothing}})
-precompile(Tuple{typeof(Base.start), Tuple{Base.Markdown.MD, Nothing}})
-precompile(Tuple{typeof(Base.indexed_next), Tuple{Base.Markdown.MD, Nothing}, Int64, Int64})
-precompile(Tuple{typeof(Base.Multimedia.display), Base.Markdown.MD})
-precompile(Tuple{typeof(Core.Compiler.eltype), Type{Array{T, 2} where T}})
-precompile(Tuple{getfield(Base.Markdown, Symbol("#kw##wrapped_lines")), Array{Any, 1}, typeof(Base.Markdown.wrapped_lines), String})
-precompile(Tuple{typeof(Base.start), Tuple{Symbol, String}})
-precompile(Tuple{typeof(Base.Markdown.terminline), Base.GenericIOBuffer{Array{UInt8, 1}}, Array{Any, 1}})
-precompile(Tuple{typeof(Base.Markdown.terminline), Base.GenericIOBuffer{Array{UInt8, 1}}, String})
-precompile(Tuple{typeof(Base.Markdown.terminline), Base.GenericIOBuffer{Array{UInt8, 1}}, Base.Markdown.Code})
-precompile(Tuple{typeof(Base._search), Base.SubString{String}, String, Int64})
-precompile(Tuple{typeof(Base._split), Base.SubString{String}, String, Int64, Bool, Array{Base.SubString{String}, 1}})
-precompile(Tuple{getfield(Base.Markdown, Symbol("#kw##wrapped_lines")), Array{Any, 1}, typeof(Base.Markdown.wrapped_lines), Base.SubString{String}})
-precompile(Tuple{typeof(Base.lastindex), Array{AbstractString, 1}})
 precompile(Tuple{typeof(Base.getindex), Array{AbstractString, 1}, Base.UnitRange{Int64}})
 precompile(Tuple{typeof(Base.throw_boundserror), Array{AbstractString, 1}, Tuple{Base.UnitRange{Int64}}})
 precompile(Tuple{typeof(Base.unsafe_copyto!), Array{AbstractString, 1}, Int64, Array{AbstractString, 1}, Int64, Int64})
