+++ stdlib/REPL/src/precompile.jl
@@ -222,15 +222,10 @@ function _precompile_()
     precompile(Tuple{typeof(Base.:(==)), Symbol, REPL.LineEdit.Prompt})
     precompile(Tuple{typeof(Base.isempty), REPL.LineEdit.PromptState})
     precompile(Tuple{typeof(REPL.LineEdit.default_enter_cb), REPL.LineEdit.PromptState})
-    precompile(Tuple{typeof(Base.Multimedia.display), REPLDisplay{REPL.LineEditREPL}, Base.Markdown.MD})
     precompile(Tuple{typeof(Base.displaysize), REPL.Terminals.TTYTerminal})
-    precompile(Tuple{typeof(Base.Markdown.term), REPL.Terminals.TTYTerminal, Array{Any, 1}, Int64})
-    precompile(Tuple{typeof(Base.Markdown.term), REPL.Terminals.TTYTerminal, Base.Markdown.MD, Int64})
-    precompile(Tuple{typeof(Base.Markdown.term), REPL.Terminals.TTYTerminal, Base.Markdown.Code, Int64})
     precompile(Tuple{typeof(Base.write), REPL.Terminals.TTYTerminal, Base.SubString{String}})
     precompile(Tuple{typeof(Base.print), REPL.Terminals.TTYTerminal, Base.SubString{String}})
     precompile(Tuple{typeof(Base.print), REPL.Terminals.TTYTerminal, Base.SubString{String}, Char})
-    precompile(Tuple{typeof(Base.Markdown.term), REPL.Terminals.TTYTerminal, Base.Markdown.Paragraph, Int64})
     precompile(Tuple{typeof(Base.Multimedia.display), REPLDisplay{REPL.LineEditREPL}, Array{Int64, 1}})
     precompile(Tuple{typeof(Base.Multimedia.display), REPLDisplay{REPL.LineEditREPL}, Base.MIME{Symbol("text/plain")}, Array{Int64, 1}})
     precompile(Tuple{typeof(Base.show_delim_array), Base.IOContext{REPL.Terminals.TTYTerminal}, Array{Int64, 1}, String, String, String, Bool, Int64, Int64})
